+++ shared/versions.ts
@@ -1,4 +1,4 @@
-export const APP_VERSION = "2.3.171";
+export const APP_VERSION = "2.3.172";
 export const ANDROID_APP_VERSION = "2.3.48";
-export const ANDROID_APK_RELEASE_VERSION = "2.3.171";
-export const AGENT_VERSION = "2.2.108";
+export const ANDROID_APK_RELEASE_VERSION = "2.3.172";
+export const AGENT_VERSION = "2.2.109";